requirements.txt
python3-flask 
python3-gpiozero
systemd-timesyncd